requirements.txt
# django packages
django-recaptcha